moved s3 uploading code from image.servie to image before find hook and replaced hard coded instances of bucket name with configuration variable
